replace ambasciata e ambasciatore with ambassador
original translation is not good - in this case using Ambassador in place of "Ambasciata" or "Ambasciatore" (not sure why two translations for the same english term) would be preferable. Additionally the image has "Ambassador" in it, even if it's been translated.
I only started the fix at the top pf the article, but please fix across all of it. Also please make a term change so to keep it in english going forward. thanks!!